N/A 패키지 정리 및 ui 공통화
