+++ feature/lesson/src/main/java/com/prography/lesson/compose/create/screen/CompleteScreen.kt
@@ -1,0 +1,15 @@
+package com.prography.lesson.compose.create.screen
+
+import androidx.compose.runtime.Composable
+import com.prography.lesson.CreateLessonIntent
+
+/**
+ * Created by MyeongKi.
+ */
+
+@Composable
+internal fun CompleteScreen(
+    intent: (CreateLessonIntent) -> Unit
+) {
+
+}
